Add mechanism for updating messages when an alloc/dep chanages.
Previously a new message would be sent everytime to state of a
deployment or allocation changed. This caused quite a lot of noise
with channels and made it hard to figure out the final outcome.

This change adds the ability to send updates to messages so that
each message will eventually represent the final outcome of that
action. To perform this a small internal state tracking system
has been added as this is needed to update Slack messages. After
this work is needed to prune the state, otherwise it will grow
and grow until the application is restarted.

Closes #1