Add region identifier to the deployment notifications.
This change adds a Nomad region CLI var for configuring the Nomad
client. In addition, if this is set the Nomad region will be added
to the notifications heading. If not passed, when setting up the
client, the local agent will be interogated to determine the local
region. Adding the region to notifications, is important in envs
which have multiple Nomad region.